+++ index.html
@@ -11,7 +11,7 @@
 
 <body>
   <main>
-    <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
